backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
supabase==2.0.3
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3 